requirements.txt
torch
transformers
datasets
numpy
scipy
sklearn
matplotlib
pyyaml
flask
logzero
sentence_transformers
ufal.morphodita
conllu
fuzzywuzzy
Levenshtein
openai
langchain
aiml
nltk
